requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
cssselect==1.2.0
html5lib==1.1
idna==3.6
lxml==5.1.0
markdownify==0.11.6
readability-lxml==0.8.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.0
webencodings==0.5.1